requirements.txt
customtkinter
openai
Pillow
pyyaml